+++ apps/ui/hooks/auth/index.ts
@@ -1,4 +1,4 @@
-import { useCurrentUser } from "../user";
+import { useCurrentUser } from "..";
 
 // TODO this should be removed because we have middleware for the check.
 // This causes a flash of unauthenticated content on page load even though the user is authenticated.
