+++ apps/ui/pages/reservations/[id]/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from "react";
+import React, { useMemo } from "react";
 import type { GetServerSidePropsContext } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import styled from "styled-components";
@@ -56,10 +56,9 @@ import {
   ButtonLikeLink,
   ButtonLikeExternalLink,
 } from "@/components/common/ButtonLikeLink";
-import { useRouter } from "next/router";
-import { successToast } from "common/src/common/toast";
 import { ReservationPageWrapper } from "@/components/reservations/styles";
 import { getReservationPath, getReservationUnitPath } from "@/modules/urls";
+import { useToastIfQueryParam } from "@/hooks";
 
 type PropsNarrowed = Exclude<Props, { notFound: boolean }>;
 
@@ -272,37 +271,6 @@ function ReservationInfo({
   );
 }
 
-function useToastIfUpdated() {
-  const router = useRouter();
-  const { t } = useTranslation();
-
-  useEffect(() => {
-    const removeTimeUpdatedParam = () => {
-      const { pathname, query } = router;
-      // NOTE ParsedQuery is a Record<string, string>
-      const params = new URLSearchParams(query as Record<string, string>);
-      params.delete("timeUpdated");
-
-      router.replace(
-        {
-          pathname,
-          query: params.toString(),
-        },
-        undefined,
-        { shallow: true }
-      );
-    };
-    const q = router.query;
-
-    if (q.timeUpdated) {
-      successToast({
-        text: t("reservations:saveNewTimeSuccess"),
-      });
-      removeTimeUpdatedParam();
-    }
-  }, [router, t]);
-}
-
 // TODO add a state check => if state is Created redirect to the reservation funnel
 // if state is Cancelled, Denied, WaitingForPayment what then?
 function Reservation({
@@ -359,7 +327,10 @@ function Reservation({
   const normalizedOrderStatus =
     getNormalizedReservationOrderStatus(reservation);
 
-  useToastIfUpdated();
+  useToastIfQueryParam({
+    key: "timeUpdated",
+    successMessage: t("reservations:saveNewTimeSuccess"),
+  });
 
   const { begin, end } = reservation;
   const timeString = capitalize(
